Used windows.confirm before delete

+++ src/containers/magasin/grid/index.js
@@ -51,17 +51,21 @@ const mapDispatchToProps = (dispatch, props) => {
     onEdit: (unit) => { hashHistory.push(`/magasin/${unit.id}/view`) },
     onDelete: (id, currentNode) => { // TODO: Problems with delete slower then callback (async)
       if (id === currentNode.id) {
-        dispatch(deleteUnit(id, {
-          onSuccess: () => {
-            dispatch(clearRoot())
-            if (currentNode.isPartOf) {
-              dispatch(loadChildren(currentNode.isPartOf))
-              dispatch(loadRoot(currentNode.isPartOf))
-            } else {
-              dispatch(loadRoot())
+        const name = currentNode.name
+        if (window.confirm(`Vil du virkelig slette node med navn ${name}`)) {
+          dispatch(deleteUnit(id, {
+            onSuccess: () => {
+              dispatch(clearRoot())
+              if (currentNode.isPartOf) {
+                dispatch(loadChildren(currentNode.isPartOf))
+                dispatch(loadRoot(currentNode.isPartOf))
+              } else {
+                dispatch(loadRoot())
+              }
+              window.alert(`Du har slettet noden med navn ${name}`)
             }
-          }
-        }))
+          }))
+        }
       }
     }
   })
